fix: レート認識を小数点以下まで対応
manga_ocrが全角文字で出力するためNFKC正規化で半角変換し、
intに丸めずfloatのまま返すよう修正。

+++ database/battle.py
@@ -18,7 +18,7 @@ class Battle:
     result: Optional[int]
     favorite: Optional[int]
     opponent_tn: Optional[str]
-    opponent_rate: Optional[int]
+    opponent_rate: Optional[float]
     battle_memo: Optional[str]
     player_party_num: Optional[int]
     player_party_subnum: Optional[int]
